Upgrade dev-deps grunt to v0.4.5; use passed version of grunt in tasks
All deprecated functionality has been removed (see:
http://gruntjs.com/api/grunt.util#external-libraries). The passed-in
grunt object for tasks is now passed into each command after the done
arg. In practice, this is not a huge issue, but it's a problem when
developing against a local grunt-git (using npm link), because tasks
would require a different version of grunt and properties could not
be set properly in the original grunt config (e.g., command_gitlog).